Create message.dart and get methods
